requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
black==22.6.0
boto3==1.24.36
botocore==1.27.36
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.6
django-admin-tools==0.9.2
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
django-minio-backend==3.3.2
django-silk==5.0.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
flower==1.2.0
gprof2dot==2021.2.21
gunicorn==20.1.0
humanize==4.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
minio==7.1.11
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
